+++ test/unit/strats.unit.test.ts
@@ -43,6 +43,9 @@ import BlastMangroveAmplifier_v2_1_0 from "../../src/assets/strats/v2.1.0/BlastM
 import BlastMangroveOrderRouter_v2_1_0 from "../../src/assets/strats/v2.1.0/BlastMangroveOrder-Router.json";
 import BlastMangroveOrder_v2_1_0 from "../../src/assets/strats/v2.1.0/BlastMangroveOrder.json";
 import BlastRouterProxyFactory_v2_1_0 from "../../src/assets/strats/v2.1.0/BlastRouterProxyFactory.json";
+import BlastKandelSeeder_v2_1_0 from "../../src/assets/strats/v2.1.0/BlastKandelSeeder.json";
+import BlastKandelLib_v2_1_0 from "../../src/assets/strats/v2.1.0/BlastKandelLib.json";
+import ZeroLendLogic_v2_1_0 from "../../src//assets/strats/v2.1.0/ZeroLendLogic.json";
 
 import {
   getAaveKandelSeederVersionDeployments,
@@ -76,6 +79,9 @@ import {
   getAllOrbitLogicVersionDeploymentsPerNetwork,
   getOrbitLogicVersionDeployments,
   getLatestOrbitLogicPerNetwork,
+  getAllZeroLendLogicVersionDeploymentsPerNetwork,
+  getLatestZeroLendLogicPerNetwork,
+  getZeroLendLogicVersionDeployments,
 } from "../../src/strats";
 import { expect } from "chai";
 import { firstVersionDeploymentsToVersionNetworkDeployment } from "./unitTestUtil";
@@ -177,11 +183,13 @@ describe("strats.ts", () => {
     describe("getKandelLibVersionDeployments", () => {
       it("should find the latest deployment first", () => {
         const result = getKandelLibVersionDeployments({ released: undefined });
-        assert.equal(result, KandelLib_v2_0_1_0);
+        assert.equal(result, BlastKandelLib_v2_1_0);
         // NB: Add older old versions here
-        [KandelLib_v2_0_0_b1_0, KandelLib_v1_0_0].forEach((version) => {
-          assert.notEqual(result, version);
-        });
+        [KandelLib_v2_0_1_0, KandelLib_v2_0_0_b1_0, KandelLib_v1_0_0].forEach(
+          (version) => {
+            assert.notEqual(result, version);
+          },
+        );
       });
     });
 
@@ -195,6 +203,7 @@ describe("strats.ts", () => {
             KandelLib_v2_0_0_b1_0,
             KandelLib_v1_0_0,
           ],
+          "81457": [BlastKandelLib_v2_1_0],
           "168587773": [KandelLib_v2_0_1_0],
         });
       });
@@ -208,6 +217,10 @@ describe("strats.ts", () => {
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
             KandelLib_v2_0_1_0,
             "80001",
+          ),
+          "81457": firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastKandelLib_v2_1_0,
+            "81457",
           ),
           "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
             KandelLib_v2_0_1_0,
@@ -224,9 +237,13 @@ describe("strats.ts", () => {
         const result = getKandelSeederVersionDeployments({
           released: undefined,
         });
-        assert.equal(result, KandelSeeder_v2_0_1_0);
+        assert.equal(result, BlastKandelSeeder_v2_1_0);
         // NB: Add older old versions here
-        [KandelSeeder_v2_0_0_b1_0, KandelSeeder_v1_0_0].forEach((version) => {
+        [
+          KandelSeeder_v2_0_1_0,
+          KandelSeeder_v2_0_0_b1_0,
+          KandelSeeder_v1_0_0,
+        ].forEach((version) => {
           assert.notEqual(result, version);
         });
       });
@@ -244,6 +261,7 @@ describe("strats.ts", () => {
             KandelSeeder_v2_0_0_b1_0,
             KandelSeeder_v1_0_0,
           ],
+          "81457": [BlastKandelSeeder_v2_1_0],
           "168587773": [KandelSeeder_v2_0_1_0],
         });
       });
@@ -257,6 +275,10 @@ describe("strats.ts", () => {
           "80001": firstVersionDeploymentsToVersionNetworkDeployment(
             KandelSeeder_v2_0_1_0,
             "80001",
+          ),
+          "81457": firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastKandelSeeder_v2_1_0,
+            "81457",
           ),
           "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
             KandelSeeder_v2_0_1_0,
@@ -482,6 +504,46 @@ describe("strats.ts", () => {
     });
   });
 
+  describe("ZeroLendLogic contract", () => {
+    describe("getZeroLendLogicVersionDeployments", () => {
+      it("should find the latest deployment first", () => {
+        const result = getZeroLendLogicVersionDeployments({
+          released: undefined,
+        });
+        assert.equal(result, ZeroLendLogic_v2_1_0);
+        // NB: Add older old versions here
+        [].forEach((version) => {
+          assert.notEqual(result, version);
+        });
+      });
+    });
+
+    describe("getAllZeroLendLogicVersionDeploymentsPerNetwork", () => {
+      it("should return all deployments grouped by network", () => {
+        expect(
+          getAllZeroLendLogicVersionDeploymentsPerNetwork({
+            released: undefined,
+          }),
+        ).to.deep.equal({
+          "81457": [ZeroLendLogic_v2_1_0],
+        });
+      });
+    });
+
+    describe("getLatestZeroLendLogicPerNetwork", () => {
+      it("should return the latest deployment for each network", () => {
+        expect(
+          getLatestZeroLendLogicPerNetwork({ released: undefined }),
+        ).to.deep.equal({
+          "81457": firstVersionDeploymentsToVersionNetworkDeployment(
+            ZeroLendLogic_v2_1_0,
+            "81457",
+          ),
+        });
+      });
+    });
+  });
+
   describe("OrbitLogic contract", () => {
     describe("getOrbitLogicVersionDeployments", () => {
       it("should find the latest deployment first", () => {
@@ -503,6 +565,7 @@ describe("strats.ts", () => {
             released: undefined,
           }),
         ).to.deep.equal({
+          "81457": [OrbitLogic_v2_1_0_1],
           "168587773": [OrbitLogic_v2_1_0_1],
         });
       });
@@ -513,6 +576,10 @@ describe("strats.ts", () => {
         expect(
           getLatestOrbitLogicPerNetwork({ released: undefined }),
         ).to.deep.equal({
+          "81457": firstVersionDeploymentsToVersionNetworkDeployment(
+            OrbitLogic_v2_1_0_1,
+            "81457",
+          ),
           "168587773": firstVersionDeploymentsToVersionNetworkDeployment(
             OrbitLogic_v2_1_0_1,
             "168587773",
@@ -624,6 +691,7 @@ describe("strats.ts", () => {
             "80001",
           ),
           orbitLogic: undefined,
+          zeroLendLogic: undefined,
         },
         "81457": {
           mangrove: firstVersionDeploymentsToVersionNetworkDeployment(
@@ -632,8 +700,14 @@ describe("strats.ts", () => {
           ),
           aaveKandelSeeder: undefined,
           aavePooledRouter: undefined,
-          kandelLib: undefined,
-          kandelSeeder: undefined,
+          kandelLib: firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastKandelLib_v2_1_0,
+            "81457",
+          ),
+          kandelSeeder: firstVersionDeploymentsToVersionNetworkDeployment(
+            BlastKandelSeeder_v2_1_0,
+            "81457",
+          ),
           mangroveOrderRouter:
             firstVersionDeploymentsToVersionNetworkDeployment(
               BlastMangroveOrderRouter_v2_1_0,
@@ -652,7 +726,14 @@ describe("strats.ts", () => {
             BlastMangroveAmplifier_v2_1_0,
             "81457",
           ),
-          orbitLogic: undefined,
+          orbitLogic: firstVersionDeploymentsToVersionNetworkDeployment(
+            OrbitLogic_v2_1_0_1,
+            "81457",
+          ),
+          zeroLendLogic: firstVersionDeploymentsToVersionNetworkDeployment(
+            ZeroLendLogic_v2_1_0,
+            "81457",
+          ),
         },
         "11155111": {
           mangrove: firstVersionDeploymentsToVersionNetworkDeployment(
@@ -669,6 +750,7 @@ describe("strats.ts", () => {
           simpleAaveLogic: undefined,
           mangroveAmplifier: undefined,
           orbitLogic: undefined,
+          zeroLendLogic: undefined,
         },
         "168587773": {
           mangrove: firstVersionDeploymentsToVersionNetworkDeployment(
@@ -707,6 +789,7 @@ describe("strats.ts", () => {
             OrbitLogic_v2_1_0_1,
             "168587773",
           ),
+          zeroLendLogic: undefined,
         },
       });
     });