+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Next version
 
+- Add zero lend adress
+
+# 2.2.1-5
+
+# 2.2.1-4
+
+# 2.2.1-3
+
+# 2.2.1-2
+
+# 2.2.1-1
+
+- Add dependencies to Blast Kandel Contracts
+
+# 2.2.1-0
+
+- Add blast kandel contract
+
 # 2.2.0
 
 - Add `OrbitLogic` contract to strats